Add `Location` component and reconstruct folder
