+++ frontend/src/ui/Component/Header.css
@@ -1,0 +1,15 @@
+.header-nav .navbar-brand {
+    display: inline-flex;
+    justify-content: center;
+    align-items: center;
+    padding: 16px 24px;
+    gap: 12px;
+}
+
+.home-nav {
+    font-size: 16px;
+    font-family: "DM", "Sans";
+    font-style: normal;
+    font-weight: 700;
+    line-height: 18px;
+}
